Remove old footer
